improve search function and cart item layout

+++ app/src/main/java/com/example/jetsnackme/model/Snack.kt
@@ -3,7 +3,7 @@ package com.example.jetsnackme.model
 data class Snack(
     val id:Int,
     val name:String,
-    val price:Double,
+    val price:Long,
     //val details:String,
     val imageUrl:String,
     val tagline:String = ""
